Move time function to utils
