+++ utils.py
@@ -1,7 +1,7 @@
 import os
 
 import discord
-
+import re
 
 async def send_dm(user, message, embed=None):
     """ Send DM to a user/member """
@@ -22,3 +22,22 @@ def is_mod(ctx):
             return True
 
     return False
+
+
+def is_valid_time(time):
+    return re.match(r"^\d+[mhd]?$", time)
+
+
+def to_minutes(time):
+    if time[-1:] == "m":
+        return int(time[:-1])
+    elif time[-1:] == "h":
+        h = int(time[:-1])
+        return h * 60
+    elif time[-1:] == "d":
+        d = int(time[:-1])
+        h = d * 24
+        return h * 60
+
+    return int(time)
+
